# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Salam Həyətimiz API
  version: 1.2.0
  summary: REST API for the Salam Həyətimiz remote access-control platform.
  description: |
    # Overview

    This document is the **binding contract** between the Laravel backend and all clients
    (Flutter mobile app, Blade admin panel, and Kapital Bank webhooks). It is the source
    of truth referenced by `TECHNICAL_SPECIFICATION.md` §9 and aligns with `DATABASE_ARCHITECTURE.md` v1.0.

    The API is split into two surfaces:

    | Surface | Prefix | Audience | Auth |
    |---|---|---|---|
    | Mobile / Public | `/v1` | End users, device owners, technical mobile mode | Phone + OTP → JWT (`bearerAuth`) |
    | Admin | `/admin/v1` | Super admin, technical (back-office) | Email + password + TOTP → JWT (`adminBearerAuth`) |
    | Webhooks | `/v1/payments/callback` | Kapital Bank | HMAC signature + IP allowlist |

    All times are **UTC** in **RFC 3339** form. All money is in **minor units** (qəpik;
    1 AZN = 100). Currency is always **AZN**.

    # Authentication

    ## Mobile (`bearerAuth`)
    Phone-based OTP authentication. The flow:

    1. `POST /auth/otp/request` with phone — backend dispatches a 6-digit SMS OTP (TTL 120 s, max 5 attempts).
    2. `POST /auth/otp/verify` with phone + code + device fingerprint — backend issues a JWT access token (15 min) and an opaque refresh token (60 days).
    3. Send `Authorization: Bearer <access_token>` on subsequent calls.
...

...

...

...

...

...

    # Authorization

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /admin/devices:
    get:
      tags: [Admin / Devices]
      summary: List devices
      operationId: adminListDevices
      security:
        - adminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - in: query
          name: status
          schema: { type: string, enum: [unassigned, active, suspended, disabled, decommissioned] }
        - in: query
          name: owner_user_id
          schema: { type: integer, format: int64 }
        - in: query
          name: region_id
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string, maxLength: 80 }
          description: Search by serial or sim_phone
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminDeviceListResponse' }
    post:
      tags: [Admin / Devices]
      summary: Register a device (alternative to mobile technical mode)
      operationId: adminCreateDevice
      security:
        - adminBearerAuth: []
      x-authorization: technical or super_admin
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceTechRegister' }
      responses:
        '201':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/devices/{deviceId}:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    get:
      tags: [Admin / Devices]
      summary: Device detail (admin)
      operationId: adminGetDevice
      security:
        - adminBearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdminDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Admin / Devices]
      summary: Update device
      operationId: adminUpdateDevice
      security:
        - adminBearerAuth: []
      x-authorization: technical or super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceAdminUpdate' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Admin / Devices]
      summary: Decommission (soft delete)
      operationId: adminDecommissionDevice
      security:
        - adminBearerAuth: []
      x-authorization: super_admin
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3, maxLength: 255 }
      responses:
        '204': { description: ok }
  /admin/devices/{deviceId}/image:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    post:
      tags: [Admin / Devices]
      summary: Upload the barrier photo (multipart)
      description: >
        Admin-only image upload. Server-side validated (real image + MIME + extension, ≤5 MB), optimised
        (downscaled to 1600px + WebP when GD is available), stored on the public disk, and the resulting
        public URL is written to `image_url`. Replaces (and deletes) any previously uploaded file. The
        response `image_url` contract is unchanged — a plain string URL.
      operationId: adminUploadDeviceImage
      security:
        - adminBearerAuth: []
      x-authorization: technical or super_admin (devices.update)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: JPG, JPEG, PNG or WEBP — max 5 MB
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Admin / Devices]
      summary: Remove the barrier photo
      operationId: adminDeleteDeviceImage
      security:
        - adminBearerAuth: []
      x-authorization: technical or super_admin (devices.update)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
...
  /admin/devices/{deviceId}/disable:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    post:
      tags: [Admin / Devices]
      summary: Disable (block opens; keep history)
      operationId: adminDisableDevice
      security:
        - adminBearerAuth: []
      x-authorization: super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, maxLength: 255 }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
  /admin/devices/{deviceId}/enable:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    post:
      tags: [Admin / Devices]
      summary: Re-enable a disabled device
      operationId: adminEnableDevice
      security:
        - adminBearerAuth: []
      x-authorization: super_admin
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
  /admin/devices/{deviceId}/transfer:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    post:
      tags: [Admin / Devices]
      summary: Transfer ownership to another user
      operationId: adminTransferDevice
      security:
        - adminBearerAuth: []
      x-authorization: super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_owner_phone, reason]
              properties:
                new_owner_phone: { type: string, pattern: '^\+994\d{9}$' }
                reason: { type: string, maxLength: 255 }
                keep_existing_users: { type: boolean, default: true }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAdmin' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Mobile JWT issued by `/auth/otp/verify` or `/auth/refresh`.
        Algorithm RS256, 15-minute TTL. Subject is `user.{id}`.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin JWT issued by `/admin/v1/auth/2fa/verify`. Disjoint identity space
        from `bearerAuth`. Algorithm RS256, 30-minute TTL.
    bankCallbackSignature:
      type: apiKey
      in: header
      name: X-Kapital-Signature
      description: |
        HMAC-SHA256 hex digest of the canonical request body using the shared
        secret negotiated with Kapital Bank. Verified before any state mutation.
    smsProviderSignature:
      type: apiKey
      in: header
      name: X-Sms-Signature
      description: HMAC signature from the inbound SMS provider.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    DeviceId:
      in: path
      name: deviceId
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    Limit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      in: query
      name: cursor
      required: false
      schema: { type: string, maxLength: 200 }
      description: Opaque pagination cursor returned in previous `page.next_cursor`.
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, minLength: 1, maxLength: 60 }
      description: |
        Optional. Same `(scope, actor, key)` returns the cached response for 24 h.
        Same key + different body → `409 idempotency_mismatch`.
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Field-level validation failures
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: subscription_required
            message_key:
              type: string
              example: errors.subscription_required
            message:
              type: string
              example: Aboneliyiniz başa çatıb.
            details:
              type: object
              additionalProperties: true
              nullable: true
            request_id:
              type: string
              example: 01JCGZ9F9XZ7P0X5MFK7Z7VYNZ
    ValidationErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            fields:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
              example:
                phone: ['Telefon nömrəsi düzgün formatda deyil.']
                code: ['Kod 6 rəqəm olmalıdır.']
            request_id: { type: string }

    # ----- Common --------------------------------------------------------
    PageInfo:
      type: object
      required: [has_more, limit]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more: { type: boolean }
        limit: { type: integer }
...
    DeviceUser:
      type: object
      required: [id, user, role, status, added_at]
      properties:
        id: { type: integer, format: int64 }
        user: { $ref: '#/components/schemas/UserBrief' }
        role: { type: string, enum: [owner, user] }
        status: { type: string, enum: [active, revoked] }
        added_by_user: { $ref: '#/components/schemas/UserBrief' }
        added_at: { type: string, format: date-time }
        last_open_at: { type: string, format: date-time, nullable: true }
        subscription: { $ref: '#/components/schemas/SubscriptionBrief' }
    UserBrief:
      type: object
      properties:
        id: { type: integer, format: int64 }
        phone_masked: { type: string, example: '+994 50 *** 45 67' }
        full_name: { type: string, nullable: true }
...
    SubscriptionBrief:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tier: { type: string, enum: [main, additional] }
        status: { type: string, enum: [pending_payment, active, expired, cancelled, refunded] }
        ends_at: { type: string, format: date-time, nullable: true }
        days_remaining: { type: integer, nullable: true }
        auto_renew: { type: boolean }
...
    DeviceTechRegister:
      type: object
      additionalProperties: false
      required: [serial, device_model_id, sim_phone, driver_type]
      properties:
        serial: { type: string, minLength: 4, maxLength: 64 }
        device_model_id: { type: integer, format: int64 }
        sim_phone: { type: string, pattern: '^\+994\d{9}$' }
        sim_operator_id: { type: integer, format: int64, nullable: true }
        sim_iccid: { type: string, pattern: '^\d{19,22}$', nullable: true }
        driver_type: { type: string, enum: [traccar, ble, sms] }
        firmware_version: { type: string, maxLength: 40, nullable: true }
        region_id: { type: integer, format: int64, nullable: true }
        location_label: { type: string, maxLength: 160, nullable: true }
        image_url: { type: string, maxLength: 2048, nullable: true }
        address: { type: string, maxLength: 255, nullable: true }
        latitude: { type: number, nullable: true, minimum: -90, maximum: 90 }
        longitude: { type: number, nullable: true, minimum: -180, maximum: 180 }
        metadata: { type: object, additionalProperties: true, nullable: true }
    DeviceAdmin:
      type: object
      properties:
        id: { type: integer, format: int64 }
        serial: { type: string }
        sim_phone: { type: string }
        sim_operator: { $ref: '#/components/schemas/SimOperator' }
        device_model: { $ref: '#/components/schemas/DeviceModel' }
        driver_type: { type: string, enum: [traccar, ble, sms] }
        status: { type: string, enum: [unassigned, active, suspended, disabled, decommissioned] }
        owner: { $ref: '#/components/schemas/UserBrief' }
        region: { $ref: '#/components/schemas/Region' }
        location_label: { type: string, nullable: true }
        image_url: { type: string, maxLength: 2048, nullable: true }
        address: { type: string, maxLength: 255, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        geofence_enabled: { type: boolean, description: '*(GEOFENCE-1)* Distance restriction toggle.' }
        geofence_radius_m: { type: integer, nullable: true, minimum: 1, maximum: 65535, description: Allowed distance in metres (null when off). }
        last_online_at: { type: string, format: date-time, nullable: true }
        last_signal_strength: { type: integer, nullable: true }
        whitelist_capacity_used: { type: integer }
        created_at: { type: string, format: date-time }
    DeviceAdminDetail:
      allOf:
        - { $ref: '#/components/schemas/DeviceAdmin' }
        - type: object
          properties:
            users:
              type: array
              items: { $ref: '#/components/schemas/DeviceUser' }
            stats:
              type: object
              properties:
                opens_30d: { type: integer }
                success_rate_30d: { type: number }
                last_open_at: { type: string, format: date-time, nullable: true }
            metadata: { type: object, additionalProperties: true }
    DeviceAdminUpdate:
      type: object
      additionalProperties: false
      properties:
        firmware_version: { type: string, maxLength: 40, nullable: true }
        driver_type: { type: string, enum: [traccar, ble, sms] }
        sim_operator_id: { type: integer, nullable: true }
        sim_iccid: { type: string, pattern: '^\d{19,22}$', nullable: true }
        region_id: { type: integer, nullable: true }
        location_label: { type: string, maxLength: 160, nullable: true }
        image_url: { type: string, maxLength: 2048, nullable: true }
        address: { type: string, maxLength: 255, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        geofence_enabled: { type: boolean, description: '*(GEOFENCE-1)* Enabling needs a positive geofence_radius_m and existing device coordinates (else 422).' }
        geofence_radius_m: { type: integer, nullable: true, minimum: 1, maximum: 65535 }
        metadata: { type: object, additionalProperties: true, nullable: true }
    AdminDeviceListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/DeviceAdmin' }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    SimOperator:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        country_iso: { type: string }
        mcc_mnc: { type: string, nullable: true }
        is_active: { type: boolean }
    DeviceModel:
      type: object
      properties:
        id: { type: integer }
        vendor: { type: string }
        model_code: { type: string }
        supports_clip: { type: boolean }
        supports_sms: { type: boolean }
        supports_mqtt: { type: boolean }
        default_driver_type: { type: string, enum: [traccar, ble, sms] }
        whitelist_capacity: { type: integer, nullable: true }
        sms_open_command: { type: string, nullable: true }
        is_active: { type: boolean }
    Region:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        parent_id: { type: integer, nullable: true }
        is_active: { type: boolean }

    # ----- Reports ------------------------------------------------------
...